refactor pet pipelines tests
